+++ src/plugin/PluginEditor.h
@@ -43,6 +43,7 @@ private:
     juce::Slider neckSlideSlider;
     juce::Slider slideFretStepsSlider;
     juce::Slider slideLiftSlider;
+    juce::Slider slideSqueakSlider;
     juce::Slider fingerNoiseSlider;
     juce::Slider ampFeedbackSlider;
     juce::Slider vibratoSpeedSlider;
@@ -68,6 +69,7 @@ private:
     juce::Label neckSlideLabel;
     juce::Label slideFretStepsLabel;
     juce::Label slideLiftLabel;
+    juce::Label slideSqueakLabel;
     juce::Label lookaheadLabel;
     juce::Label fingerNoiseLabel;
     juce::Label ampFeedbackLabel;
@@ -103,6 +105,7 @@ private:
     juce::TextButton neckSlideInfoButton;
     juce::TextButton slideFretStepsInfoButton;
     juce::TextButton slideLiftInfoButton;
+    juce::TextButton slideSqueakInfoButton;
     juce::TextButton lookaheadInfoButton;
     juce::TextButton fingerNoiseInfoButton;
     juce::TextButton ampFeedbackInfoButton;
@@ -140,6 +143,7 @@ private:
     std::unique_ptr<SliderAttachment> neckSlideAttachment;
     std::unique_ptr<SliderAttachment> slideFretStepsAttachment;
     std::unique_ptr<SliderAttachment> slideLiftAttachment;
+    std::unique_ptr<SliderAttachment> slideSqueakAttachment;
     std::unique_ptr<ComboBoxAttachment> lookaheadAttachment;
     std::unique_ptr<SliderAttachment> fingerNoiseAttachment;
     std::unique_ptr<SliderAttachment> ampFeedbackAttachment;